Extract entities into their own files
