+++ src/domain/aggregated/entities/AggregatedPackage.ts
@@ -1,0 +1,5 @@
+import { DataValue } from "./DataValue";
+
+export interface AggregatedPackage {
+    dataValues: DataValue[];
+}
