+++ src/domain/events/entities/EventsPackage.ts
@@ -1,0 +1,5 @@
+import { ProgramEvent } from "./ProgramEvent";
+
+export interface EventsPackage {
+    events: ProgramEvent[];
+}
