+++ src/domain/events/entities/ProgramEvent.ts
@@ -1,7 +1,4 @@
-export interface EventsPackage {
-    events: ProgramEvent[];
-}
-
+import { ProgramEventDataValue } from "./ProgramEventDataValue";
 export interface ProgramEvent {
     id: string;
     orgUnit: string;
@@ -19,13 +16,4 @@ export interface ProgramEvent {
     attributeCategoryOptions?: string;
     attributeOptionCombo?: string;
     dataValues: ProgramEventDataValue[];
-}
-
-export interface ProgramEventDataValue {
-    lastUpdated: string;
-    storedBy: string;
-    created: string;
-    dataElement: string;
-    value: any;
-    providedElsewhere: boolean;
 }
