+++ src/domain/events/entities/ProgramEventDataValue.ts
@@ -1,0 +1,8 @@
+export interface ProgramEventDataValue {
+    lastUpdated: string;
+    storedBy: string;
+    created: string;
+    dataElement: string;
+    value: any;
+    providedElsewhere: boolean;
+}
