+++ src/domain/events/repositories/EventsRepository.ts
@@ -1,5 +1,5 @@
 import { DataSynchronizationParams } from "../../aggregated/types";
-import { ProgramEvent } from "../entities/Events";
+import { ProgramEvent } from "../entities/ProgramEvent";
 import { DataImportParams, DataImportResponse } from "../../../types/d2";
 
 export interface EventsRepository {
